# src/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: Shopix E-Commerce API
  version: "1.0.0"
  description: |
    Full REST API documentation for the **Shopix** e-commerce platform.

    ### Authentication
    Most endpoints require a valid `Accesstoken` cookie set after login.
    Pass credentials via the login endpoint — the server sets HttpOnly cookies automatically.

  contact:
    name: Shopix Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: Accesstoken
  schemas:
...
    Order:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        storeId:
          type: string
        productId:
          type: string
        receiverName:
          type: string
        mobileNumber:
          type: string
        billingAddress:
          type: string
        quantity:
          type: integer
        totalAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, stripe]
        paymentStatus:
          type: string
          enum: [Pending, Paid]
        status:
          type: string
          enum: [new, processing, shipped, completed, cancelled]
        returnStatus:
          type: string
          enum: [none, processing, successful, failed]
        createdAt:
          type: string
          format: date-time

    # ─── Notification ─────────────────────────────────────────
...
paths:
...
  /api/products/{id}:
    put:
      tags: [Products]
      summary: Update a product
      description: Updates fields of an existing product. Only the owning seller can update. Accepts `multipart/form-data`; image field is optional.
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                  maxLength: 25
                price:
                  type: number
                quantity:
                  type: integer
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Product updated
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Product not found

    delete:
      tags: [Products]
      summary: Delete a product (or all products)
      description: |
        - Pass a valid **MongoDB ID** to delete a single product.
        - Pass `all` as the ID to delete **all** products in the seller's store (requires password confirmation in body).
        - Admins can delete any product by ID without password.
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Product MongoDB ID, or the literal string `all`
      requestBody:
        description: Required only when id is `all`
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Product(s) deleted
        "400":
          description: Password required (for bulk delete)
        "401":
          description: Incorrect password / Unauthorized
        "404":
          description: Product or store not found

  # ───────────────────────────────────────────────────────────
  # CART
  # ───────────────────────────────────────────────────────────
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Place a new order
      description: Creates an order for a product, deducts inventory, sends email + in-app notifications to buyer and seller.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, receiverName, mobileNumber, billingAddress, quantity, paymentMethod]
              properties:
                productId:
                  type: string
                receiverName:
                  type: string
                mobileNumber:
                  type: string
                billingAddress:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
                paymentMethod:
                  type: string
                  enum: [cash, stripe]
                stripePaymentId:
                  type: string
                  description: Required for stripe payments
                paymentStatus:
                  type: string
                  enum: [Pending, Paid]
      responses:
        "200":
          description: Order placed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  order:
                    $ref: "#/components/schemas/Order"
        "400":
          description: Missing fields or insufficient stock
        "404":
          description: Product not found

    get:
      tags: [Orders]
      summary: Get seller's store orders
      description: Returns all orders for the authenticated seller's store (most recent first).
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Seller orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
        "404":
          description: Store not found

    patch:
      tags: [Orders]
      summary: Update order status (seller)
      description: Allows the seller to update the status of an order. Triggers status email + in-app notification to buyer when status is `shipped`, `completed`, or `delivered`.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, status]
              properties:
                orderId:
                  type: string
                status:
                  type: string
                  enum: [processing, shipped, completed, cancelled]
      responses:
        "200":
          description: Order status updated
        "400":
          description: orderId and status required
        "403":
          description: Unauthorized — not seller's order
        "404":
          description: Order not found
...